Add sysreset method for entering application
Running main() of the application does not work reliably depending on its
clock configuration. An easy way to circumvent this is to reset the core
instead of using jump_to_flash(). After the CPU reset the bootloader
will run again and execute the jump to the application.
This feature can be enabled with the define USE_CORE_RESET.